refactor: use existing const for type and mimetypes
