+++ src/Provider/MimeTypesProviderInterface.php
@@ -15,5 +15,43 @@ namespace MonsieurBiz\SyliusMediaManagerPlugin\Provider;
 
 interface MimeTypesProviderInterface
 {
+    public const IMAGE_TYPE_MIMES = [
+        'image/gif',
+        'image/jpeg',
+        'image/png',
+        'image/svg+xml',
+        'image/webp',
+        'image/avif',
+    ];
+
+    public const VIDEO_TYPE_MIMES = [
+        'video/mp4',
+        'video/mov',
+        'video/3gp',
+        'video/ogg',
+        'video/webm',
+    ];
+
+    public const PDF_TYPE_MIMES = [
+        'application/pdf',
+    ];
+
+    public const FAVICON_TYPE_MIMES = [
+        'image/vnd.microsoft.icon',
+        'image/x-icon',
+        'image/ico',
+        'image/svg+xml',
+        'image/gif',
+        'image/jpeg',
+        'image/png',
+    ];
+
+    public const AUDIO_TYPE_MIMES = [
+        'audio/mpeg',
+        'audio/mpeg3',
+        'audio/x-mpeg',
+        'audio/x-mpeg-3',
+    ];
+
     public function getMimeTypesByType(?string $type): array;
 }
